+++ ReactNativeTraining/src/redux/chat/actions.js
@@ -1,3 +1,5 @@
-export const chatSelected = id => dispatch => dispatch({ type: 'CHAT_SELECTED', payload: id });
+export const chatSelected = id => dispatch => {
+  dispatch({ type: 'CHAT_SELECTED', payload: id });
+};
 
 export default chatSelected;
