+++ ReactNativeTraining/src/redux/chat/reducer.js
@@ -1,7 +1,11 @@
 export const chat = (state = [], action) => {
   switch (action.type) {
     case 'CHAT_SELECTED':
-      return action.payload;
+      return { senderId: action.payload, groupId: null };
+    case 'GROUP_SELECTED':
+      return { groupId: action.payload, senderId: null };
+    case 'CONTACT_SELECTED':
+      return { senderId: action.payload, groupId: null };
     default:
       return state;
   }
