+++ ReactNativeTraining/src/redux/contacts/actions.js
@@ -1,0 +1,5 @@
+export const contactSelected = id => dispatch => {
+  dispatch({ type: 'CONTACT_SELECTED', payload: id });
+};
+
+export default contactSelected;
