+++ ReactNativeTraining/src/redux/groups/actions.js
@@ -1,0 +1,3 @@
+export const groupSelected = id => dispatch => dispatch({ type: 'GROUP_SELECTED', payload: id });
+
+export default groupSelected;
