+++ ReactNativeTraining/src/screens/components/Conversation/styles.js
