+++ ReactNativeTraining/src/screens/components/Message/index.js
